cleanup - move flags towards call site
